Login Edited and Design Changes

+++ src/Components/Admin/ManageProduct.js
@@ -1,7 +1,8 @@
 import { Button } from '@material-ui/core';
-import React, { useEffect, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import { useHistory, useLocation } from 'react-router';
 import AdminBar from './AdminBar';
+import { UserContext } from '../../App';
 
 const ManageProduct = () => {
 
@@ -9,47 +10,79 @@ const ManageProduct = () => {
         loading: false,
         data: null,
     });
-
+    const [loggedInUser, setLoggedInUser] = useContext(UserContext);
     const history = useHistory();
     const location = useLocation();
     const { from } = location.state || { from: { pathname: "/" } };
 
-    useEffect(()=>{
+    useEffect(() => {
+        if ((sessionStorage.getItem('user'))) {
+            setLoggedInUser(JSON.parse(sessionStorage.getItem('user')));
+        }
+    }, [])
+
+    useEffect(() => {
         setProducts({ loading: true });
         fetch('https://cherry-pie-50881.herokuapp.com/products')
-        .then(res => res.json())
-        .then(data => {
-            console.log(data);
-            setProducts({ loading: false, data:data });
-        })
-        .catch(err => console.log(err));
-    },[setProducts])
+            .then(res => res.json())
+            .then(data => {
+                console.log(data);
+                setProducts({ loading: false, data: data });
+            })
+            .catch(err => console.log(err));
+    }, [setProducts])
 
     const handleDelete = (id) => {
-        console.log(id);
-        fetch('https://cherry-pie-50881.herokuapp.com/delete/' + id ,{
-            method:'DELETE'
-        })
-        .then(res => res.json())
-        .then(result => {
-            console.log(result, 'Deleted');
-            alert('Product Deleted');
-            history.replace(from);
-        })
+        if (window.confirm('Are You Sure, You want to Delete The Product?')) {
+            console.log(id);
+            if (loggedInUser.admin === true) {
+                fetch('https://cherry-pie-50881.herokuapp.com/delete/' + id, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(result => {
+                        console.log(result, 'Deleted');
+                        alert('Product Deleted');
+                        // history.replace(from);
+                    })
+            }
+            else {
+                alert("Sorry ! You are Not Authorized to Delete")
+            }
+
+        }
     }
 
     return (
-        <div style={{marginTop:'15rem'}}>
-            <div style={{marginLeft:'20%'}}>
+        <div style={{ marginTop: '15rem' }}>
+            <div style={{ marginLeft: '20%' }}>
                 <AdminBar></AdminBar>
                 <h3>Manage Product</h3>
                 {
                     products.loading &&
                     <h3>Loading Products</h3>
                 }
-                {
-                    products.data&&products.data.map(product => <li style={{padding:'10px'}}>Name:{product.name} Price:  {product.price} <Button  size="small" variant="contained" color="secondary" onClick={()=>handleDelete(product._id)}>Delete </Button> </li> )
-                }
+                <table style={{ padding: '5vh' }}>
+                    <thead style={{ padding: '5vh', margin: '5vh' }}>
+                        <tr style={{ padding: '5vh', margin: '5vh' }}>
+                            <td>Name : </td>
+                            <td>Price: </td>
+                            <td>Delete: </td>
+                        </tr>
+
+                    </thead>
+                    <tbody style={{ padding: '5vh', margin: '5vh' }}>
+                        {
+                            products.data && products.data.map(product =>
+                                <tr style={{ padding: '5vh', margin: '5vh' }}>
+                                    <td>{product.name}</td>
+                                    <td>{product.price}</td>
+                                    <td><Button size="small" variant="contained" color="secondary" onClick={() => handleDelete(product._id)}>Delete </Button></td>
+
+                                </tr>)
+                        }
+                    </tbody>
+                </table>
             </div>
         </div>
     );
